Add CoordinateFunctionFeatureExtractor; Del casts
